requirements.txt
asgiref==3.4.1
Django==4.0.1
sqlparse==0.4.2
tzdata==2021.5
PyJWT==2.3.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mysql==0.0.3
mysqlclient==2.1.0
pip==21.1.2
pytz==2021.3
setuptools==57.0.0
wheel==0.36.2
